15f

+++ src/main/java/com/example/thebra/striperequest/StripeRequest.java
@@ -10,11 +10,11 @@ import java.util.UUID;
 public class StripeRequest {
 
 
-    private Long amount;
+    private Double amount;
     private String currency;
     private String stripeToken;
     private UUID orderId;
-    public StripeRequest(Long amount, String currency, String stripeToken, UUID orderId){
+    public StripeRequest(Double amount, String currency, String stripeToken, UUID orderId){
         this.amount = amount;
         this.currency = currency;
         this.stripeToken = stripeToken;
